pyproject.toml
[project]
name = "libqcd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "sagemaker>=2.239.0",
]
